Move Fake Database Exception tests out of Version tests
